refactor: apps structure completly refactored for better scale management
