+++ src/apps/demo/core/business/index.ts
@@ -1,0 +1,2 @@
+export * from './repositories';
+export * from './services';
