+++ src/apps/demo/core/business/services/todo.service.ts
@@ -1,6 +1,5 @@
-import { ITodoModel } from '../types';
 import { TodoRepository } from '../repositories';
-import { TodoModel } from '../models';
+import { TodoModel } from '../../domain/models';
 import { parseSortParam } from 'helpers';
 import {
   ErrorResponseType,
