+++ src/apps/demo/core/domain/index.ts
@@ -1,0 +1,2 @@
+export * from './models';
+export * from './types';
