+++ src/apps/demo/core/domain/types/index.ts
