+++ src/apps/demo/index.ts
@@ -1,5 +1,1 @@
-/**
- * The only thing that need to be export here is the router
- */
-
-export * from './routes';
+export * from './core';
